+++ src/ua/data_type_definition.rs
@@ -1,37 +1,38 @@
 use open62541_sys::{UA_NS0ID_ENUMDEFINITION, UA_NS0ID_STRUCTUREDEFINITION};
 
-use crate::{DataTypeExt, ua};
+use crate::{DataTypeExt, Error, Result, ua};
 
 #[derive(Debug, Clone)]
 #[non_exhaustive]
 pub enum DataTypeDefinition {
     Structure(ua::StructureDefinition),
     Enum(ua::EnumDefinition),
-    Unknown(ua::Variant),
 }
 
 impl DataTypeDefinition {
-    pub(crate) fn new(value: ua::Variant) -> Self {
+    pub(crate) fn from_abstract(value: ua::Variant) -> Result<Self> {
         let Some(type_id) = value.type_id() else {
-            return Self::Unknown(value);
+            return Err(Error::Internal("require type ID for data type definition"));
         };
 
-        if *type_id == ua::NodeId::ns0(UA_NS0ID_STRUCTUREDEFINITION) {
-            // PANIC: We have checked that the expected type ID is set.
-            Self::Structure(value.into_scalar().expect("require structure definition"))
-        } else if *type_id == ua::NodeId::ns0(UA_NS0ID_ENUMDEFINITION) {
-            // PANIC: We have checked that the expected type ID is set.
-            Self::Enum(value.into_scalar().expect("require enum definition"))
-        } else {
-            Self::Unknown(value)
-        }
+        Ok(
+            if *type_id == ua::NodeId::ns0(UA_NS0ID_STRUCTUREDEFINITION) {
+                // PANIC: We have checked that the expected type ID is set.
+                Self::Structure(value.into_scalar().expect("structure definition"))
+            } else if *type_id == ua::NodeId::ns0(UA_NS0ID_ENUMDEFINITION) {
+                // PANIC: We have checked that the expected type ID is set.
+                Self::Enum(value.into_scalar().expect("enum definition"))
+            } else {
+                return Err(Error::Internal("unsupported data type definition"));
+            },
+        )
     }
 
+    #[must_use]
     pub(crate) fn into_abstract(self) -> ua::Variant {
         match self {
             Self::Structure(value) => ua::Variant::scalar(value),
             Self::Enum(value) => ua::Variant::scalar(value),
-            Self::Unknown(value) => value,
         }
     }
 }
@@ -39,8 +40,8 @@ impl DataTypeDefinition {
 impl DataTypeExt for DataTypeDefinition {
     type Inner = ua::Variant;
 
-    fn from_inner(value: Self::Inner) -> Self {
-        Self::new(value)
+    fn from_inner(value: Self::Inner) -> Result<Self> {
+        Self::from_abstract(value)
     }
 
     fn into_inner(self) -> Self::Inner {