+++ src/ua/data_types/string.rs
@@ -42,9 +42,9 @@ impl String {
         str.into_string()
     }
 
-    /// Creates invalid string (as defined by OPC UA).
-    // TODO: The OPC UA specification calls invalid strings "null". Consider changing this to match.
-    pub(crate) fn invalid() -> Self {
+    /// Creates an invalid null string (as defined by OPC UA).
+    #[must_use]
+    pub fn null() -> Self {
         let str = unsafe { UA_String_fromChars(ptr::null()) };
         Self(str)
     }
@@ -71,12 +71,12 @@ impl String {
         }
     }
 
-    /// Checks if string is invalid.
+    /// Checks if string is null and invalid.
     ///
-    /// The invalid state is defined by OPC UA. It is a third state which is distinct from empty and
+    /// The null state is defined by OPC UA. It is a third state which is distinct from empty and
     /// regular (non-empty) strings.
     #[must_use]
-    pub fn is_invalid(&self) -> bool {
+    pub fn is_null(&self) -> bool {
         matches!(self.array_value(), ArrayValue::Invalid)
     }
 
@@ -182,8 +182,19 @@ mod tests {
     #[test]
     fn valid_string() {
         let str = ua::String::new("lorem ipsum").expect("should parse string");
+        assert!(!str.is_null());
+        assert!(!str.is_empty());
         assert_eq!(str.as_str().expect("should display string"), "lorem ipsum");
         assert_eq!(str.to_string(), "lorem ipsum");
+    }
+
+    #[test]
+    fn null_string() {
+        let str = ua::String::null();
+        assert!(str.is_null());
+        assert!(!str.is_empty());
+        assert!(str.as_str().is_none());
+        assert_eq!(str.to_string(), "");
     }
 
     #[test]
@@ -191,6 +202,8 @@ mod tests {
         // Empty strings may have an internal representation in `UA_String` that contains invalid or
         // null pointers. This must not cause any problems.
         let str = ua::String::new("").expect("should parse empty string");
+        assert!(!str.is_null());
+        assert!(str.is_empty());
         assert_eq!(str.as_str().expect("should display empty string"), "");
         assert_eq!(str.to_string(), "");
     }