+++ src/ua/data_types/structure_description.rs
@@ -1,0 +1,34 @@
+use open62541_sys::UA_StructureDescription;
+
+use crate::{DataType as _, ua};
+
+crate::data_type!(StructureDescription);
+
+impl StructureDescription {
+    pub(crate) fn new(
+        data_type_id: ua::NodeId,
+        name: ua::QualifiedName,
+        definition: ua::StructureDefinition,
+    ) -> Self {
+        Self(UA_StructureDescription {
+            dataTypeId: data_type_id.into_raw(),
+            name: name.into_raw(),
+            structureDefinition: definition.into_raw(),
+        })
+    }
+
+    #[must_use]
+    pub fn data_type_id(&self) -> &ua::NodeId {
+        ua::NodeId::raw_ref(&self.0.dataTypeId)
+    }
+
+    #[must_use]
+    pub fn name(&self) -> &ua::QualifiedName {
+        ua::QualifiedName::raw_ref(&self.0.name)
+    }
+
+    #[must_use]
+    pub fn structure_definition(&self) -> &ua::StructureDefinition {
+        ua::StructureDefinition::raw_ref(&self.0.structureDefinition)
+    }
+}