+++ src/ua/data_types/structure_type.rs
@@ -1,0 +1,13 @@
+crate::data_type!(StructureType);
+
+crate::enum_variants!(
+    StructureType,
+    UA_StructureType,
+    [
+        STRUCTURE,
+        STRUCTUREWITHOPTIONALFIELDS,
+        UNION,
+        STRUCTUREWITHSUBTYPEDVALUES,
+        UNIONWITHSUBTYPEDVALUES,
+    ],
+);